Extract `GOVCERT` ExecSpecRule and SpecTranslate instances
