+++ libs/cardano-ledger-conformance/cardano-ledger-conformance.cabal
@@ -36,6 +36,7 @@ library
         Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Cert
         Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Certs
         Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Gov
+        Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.GovCert
         Test.Cardano.Ledger.Conformance.ExecSpecRule.Core
         Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway
         Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Base
@@ -44,6 +45,7 @@ library
         Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Cert
         Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Certs
         Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Gov
+        Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.GovCert
 
     default-language: Haskell2010
     ghc-options:
