+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/ExecSpecRule/Conway.hs
@@ -5,4 +5,5 @@ import Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Cert ()
 import Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Certs ()
 import Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Deleg ()
 import Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Gov ()
+import Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.GovCert ()
 import Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Pool ()
